# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Afrik'Fid API"
  version: "1.0.0"
  description: |
    Passerelle de paiement B2B multi-pays avec fidélisation intégrée pour l'Afrique francophone et anglophone.

    ## Authentification

    L'API supporte deux mécanismes d'authentification :

    **1. JWT Bearer** — Pour le tableau de bord admin et marchand
    ```
    Authorization: Bearer <accessToken>
    ```
    Obtenir un token via `POST /auth/admin/login` ou `POST /auth/merchant/login`.

    **2. API Key** — Pour les intégrations marchand (paiements, confirmations, statuts)
    ```
    X-API-Key: af_pub_xxxxxxxxxxxx
    ```
    Ajouter `X-Sandbox: true` pour utiliser les clés sandbox et simuler les paiements.

    ## Modèle de fidélité X/Y/Z

    Chaque paiement est réparti entre trois parties :
    - **X%** — Commission marchand (remise accordée au client ou gardée)
    - **Y%** — Rebate client (selon statut loyalty : OPEN=0%, LIVE=5%, GOLD=8%, ROYAL=10%)
    - **Z%** — Commission plateforme Afrik'Fid

    ## Environnements

    | Environnement | Base URL | Clé |
    |---|---|---|
    | Production | `https://api.afrikfid.com/api/v1` | `af_pub_xxx` |
    | Sandbox | `https://api.afrikfid.com/api/v1` + `X-Sandbox: true` | `af_sandbox_pub_xxx` |
    | Local | `http://localhost:4001/api/v1` | Clé de test |

  contact:
    name: "Support Afrik'Fid"
    email: support@afrikfid.com
  license:
    name: Propriétaire
servers:
  - url: http://localhost:4001/api/v1
...
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/admin/login ou /auth/merchant/login
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API publique du marchand (af_pub_xxx en prod, af_sandbox_pub_xxx en sandbox)
    HmacSignature:
      type: apiKey
      in: header
      name: X-AfrikFid-Signature
      description: |
        Signature HMAC-SHA256 optionnelle pour l'intégrité des requêtes (CDC §4.5.2).
        Format : `sha256=<hex>`
        Calcul : `HMAC-SHA256(api_key_secret, raw_request_body)`
        En mode `REQUIRE_HMAC_SIGNATURE=true`, ce header devient obligatoire sur `/payments/initiate` et `/payments/{id}/refund`.

  # ─── Schemas ───────────────────────────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Données invalides"
      required: [error]
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
...
    Merchant:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: "Boutique Koffi" }
        email: { type: string, format: email }
        phone: { type: string, example: "+2250701000001" }
        countryId: { type: string, example: CI }
        countryName: { type: string, example: "Côte d'Ivoire" }
        currency: { type: string, example: XOF }
        category: { type: string, example: retail }
        rebatePercent: { type: number, example: 8 }
        rebateMode: { type: string, enum: [cashback, immediate] }
        settlementFrequency: { type: string, enum: [daily, weekly, monthly] }
        webhookUrl: { type: string, format: uri, nullable: true }
        status: { type: string, enum: [pending, active, suspended] }
        kycStatus: { type: string, enum: [pending, approved, rejected] }
        isActive: { type: integer, enum: [0, 1] }
        apiKeyPublic: { type: string, example: "af_pub_abc123" }
        sandboxKeyPublic: { type: string, example: "af_sandbox_pub_abc123" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Client:
      type: object
      properties:
        id: { type: string, format: uuid }
        afrikfidId: { type: string, example: "AFD-CLI-000001" }
        fullName: { type: string, example: "Jean Dupont" }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, example: "+2250700123456" }
        countryId: { type: string, example: CI }
        loyaltyStatus: { type: string, enum: [OPEN, LIVE, GOLD, ROYAL] }
        clientRebatePercent: { type: number, example: 5 }
        walletBalance: { type: number, example: 2500 }
        totalPurchases: { type: integer }
        totalAmount: { type: number }
        isActive: { type: integer, enum: [0, 1] }
        createdAt: { type: string, format: date-time }
    PaymentLink:
      type: object
      properties:
        id: { type: string, format: uuid }
        merchantId: { type: string, format: uuid }
        code: { type: string, example: "PL-ABCD-1234" }
        amount: { type: number, nullable: true, example: 5000 }
        currency: { type: string, example: XOF }
        description: { type: string, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        maxUses: { type: integer, example: 1 }
        usesCount: { type: integer, example: 0 }
        status: { type: string, enum: [active, used, cancelled, expired] }
        createdAt: { type: string, format: date-time }
    WebhookEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        merchantId: { type: string, format: uuid }
        eventType:
          type: string
          enum: [payment.completed, payment.failed, payment.refunded, loyalty.status_upgraded, webhook.test]
        url: { type: string, format: uri }
        status: { type: string, enum: [pending, delivered, failed, retry] }
        attempts: { type: integer, example: 1 }
        lastResponseCode: { type: integer, nullable: true, example: 200 }
        lastError: { type: string, nullable: true }
        nextRetryAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        sentAt: { type: string, format: date-time, nullable: true }
...
  responses:
    Unauthorized:
      description: Token manquant ou invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: "TOKEN_INVALID", message: "Token invalide" }
    Forbidden:
      description: Permissions insuffisantes
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ValidationError:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }

# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /payment-links:
    post:
      tags: [Payment Links]
      summary: Créer un lien de paiement
      description: Crée un lien partageable permettant à un client de payer sans intégration technique.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  description: Montant fixe (omis = montant libre, le client saisit)
                  example: 5000
                currency: { type: string, example: XOF }
                description: { type: string, example: "Prestation conseil" }
                expires_in_hours:
                  type: integer
                  description: Durée de validité en heures (défaut 24h)
                  example: 48
                max_uses:
                  type: integer
                  description: Nombre d'utilisations max (défaut 1)
      responses:
        '201':
          description: Lien créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  link: { $ref: '#/components/schemas/PaymentLink' }
                  payUrl:
                    type: string
                    format: uri
                    description: URL publique à partager
                    example: "https://afrikfid.com/pay/PL-ABCD-1234"
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [Payment Links]
      summary: Lister les liens du marchand connecté
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des liens
          content:
            application/json:
              schema:
                type: object
                properties:
                  links:
                    type: array
                    items: { $ref: '#/components/schemas/PaymentLink' }
...
  /merchants:
    post:
      tags: [Merchants]
      summary: Créer un marchand (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, country_id, rebate_percent, rebate_mode]
              properties:
                name: { type: string, example: "Boutique Koffi" }
                email: { type: string, format: email }
                phone: { type: string }
                country_id: { type: string, example: CI }
                category: { type: string, example: retail }
                rebate_percent:
                  type: number
                  minimum: 0
                  maximum: 50
                  example: 8
                rebate_mode: { type: string, enum: [cashback, immediate] }
                password: { type: string, description: "Mot de passe dashboard marchand" }
                webhook_url: { type: string, format: uri }
                settlement_frequency: { type: string, enum: [daily, weekly, monthly] }
      responses:
        '201':
          description: Marchand créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant: { $ref: '#/components/schemas/Merchant' }
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Email déjà utilisé

    get:
      tags: [Merchants]
      summary: Lister les marchands (admin, paginé)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: status
          schema: { type: string, enum: [pending, active, suspended] }
        - in: query
          name: country_id
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
          description: Recherche par nom ou email
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchants:
                    type: array
                    items: { $ref: '#/components/schemas/Merchant' }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
...
  /clients:
    post:
      tags: [Clients]
      summary: Créer un client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, full_name]
              properties:
                phone: { type: string, example: "+2250700123456" }
                full_name: { type: string, example: "Jean Dupont" }
                email: { type: string, format: email }
                country_id: { type: string, example: CI }
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  client: { $ref: '#/components/schemas/Client' }
        '409':
          description: Numéro de téléphone déjà utilisé

    get:
      tags: [Clients]
      summary: Lister les clients (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
          description: Recherche par nom, téléphone ou afrikfid_id
        - in: query
          name: loyalty_status
          schema: { type: string, enum: [OPEN, LIVE, GOLD, ROYAL] }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  clients:
                    type: array
                    items: { $ref: '#/components/schemas/Client' }
                  total: { type: integer }
...
  /webhooks:
    get:
      tags: [Webhooks]
      summary: Lister les événements webhook (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, delivered, failed, retry] }
        - in: query
          name: merchant_id
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/WebhookEvent' }
                  total: { type: integer }
...
  /audit-logs:
    get:
      tags: [System]
      summary: Journal d'audit complet (admin — CDC §4.6.1)
      description: |
        Retourne l'historique complet des actions sensibles de la plateforme.
        Accessible uniquement aux administrateurs Afrik'Fid.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: action
          schema: { type: string }
          description: Filtrer par type d'action (ex. kyc_approved, refund_initiated)
        - in: query
          name: actor_id
          schema: { type: string, format: uuid }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Liste des entrées d'audit
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        action: { type: string }
                        actor_id: { type: string }
                        actor_role: { type: string, enum: [admin, merchant, system] }
                        target_type: { type: string }
                        target_id: { type: string }
                        details: { type: object }
                        ip_address: { type: string }
                        created_at: { type: string, format: date-time }
                  total: { type: integer }
                  page: { type: integer }
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── Reports — PDF & Excel ──────────────────────────────────────────────────
...